+++ stable/sck.cpp
@@ -43,8 +43,6 @@ typedef int ssize_t;
 bool sck::Ready_ = true;
 #endif
 
-#include "cio.h"
-
 flw::size__ sck::Read(
 	socket__ Socket,
 	flw::size__ Amount,
@@ -72,7 +70,6 @@ flw::size__ sck::Read(
 		Result = recv( Socket, (cast__)Buffer, (int)Amount, 0 );
 
 		if ( Result == SCK_SOCKET_ERROR ) {
-      CPq;
 			Result = SCK_DISCONNECTED;
 
 			if ( Error() != SCK_ECONNRESET )
@@ -119,7 +116,6 @@ flw::size__ sck::Write(
 
 		if ( Result == SCK_SOCKET_ERROR )
 		{
-		  CPq;
 			Result = SCK_DISCONNECTED;
 
 			if ( Error() != SCK_ECONNRESET )