+++ tools/xdhq/Atlas/PYH/Atlas.py
@@ -120,7 +120,7 @@ def worker(userCallback,dom,callbacks):
 			else:
 				dom.alert("\tDEV ERROR: missing callback for '" + action + "' action!") 
 	
-	l() # Exiting thread.
+	# l() # Exiting thread, closing correspondig instance.
 
 def _callback(userCallback,callbacks,instance):
 	thread = Thread(target=worker, args=(userCallback, XDHq.DOM(instance), callbacks))